requirements.txt
fastapi==0.116.1
python-dotenv==1.1.1
Requests==2.32.4
uvicorn[standard]
python-multipart
